replace unsigned int timestamps with int64_t (#11)

+++ include/DAQInterface.h
@@ -29,21 +29,21 @@ namespace ToolFramework {
     bool SQLQuery(const std::string& database, const std::string& query, std::string& response, const unsigned int timeout=300);
     bool SQLQuery(const std::string& database, const std::string& query, const unsigned int timeout=300);
     
-    bool SendLog(const std::string& message, unsigned int severity=2, const std::string& device="", const unsigned int timestamp=0); //serverity levels are 0 = critical, 1 = Error, 2 = warning, 3= info , 4-9 debug
-    bool SendAlarm(const std::string& message, unsigned int level=0, const std::string& device="", const unsigned int timestamp=0, const unsigned int timeout=300);
-    bool SendMonitoringData(const std::string& json_data, const std::string& device="", unsigned int timestamp=0);
-    bool SendCalibrationData(const std::string& json_data, const std::string& description, const std::string& device="", unsigned int timestamp=0, int* version=nullptr, const unsigned int timeout=300);
+    bool SendLog(const std::string& message, unsigned int severity=2, const std::string& device="", const int64_t timestamp=0); //serverity levels are 0 = critical, 1 = Error, 2 = warning, 3= info , 4-9 debug
+    bool SendAlarm(const std::string& message, unsigned int level=0, const std::string& device="", const int64_t timestamp=0, const unsigned int timeout=300);
+    bool SendMonitoringData(const std::string& json_data, const std::string& device="", int64_t timestamp=0);
+    bool SendCalibrationData(const std::string& json_data, const std::string& description, const std::string& device="", int64_t timestamp=0, int* version=nullptr, const unsigned int timeout=300);
     bool GetCalibrationData(std::string& json_data, int version=-1, const std::string& device="", const unsigned int timeout=300);
-    bool SendDeviceConfig(const std::string& json_data, const std::string& author, const std::string& description, const std::string& device="", unsigned int timestamp=0, int* version=nullptr, const unsigned int timeout=300);
-    bool SendRunConfig(const std::string& json_data, const std::string& name, const std::string& author, const std::string& description, unsigned int timestamp=0, int* version=nullptr, const unsigned int timeout=300);
+    bool SendDeviceConfig(const std::string& json_data, const std::string& author, const std::string& description, const std::string& device="", int64_t timestamp=0, int* version=nullptr, const unsigned int timeout=300);
+    bool SendRunConfig(const std::string& json_data, const std::string& name, const std::string& author, const std::string& description, int64_t timestamp=0, int* version=nullptr, const unsigned int timeout=300);
     bool GetDeviceConfig(std::string& json_data, const int version, const std::string& device="", const unsigned int timeout=300);
     bool GetRunConfig(std::string& json_data, const int config_id, const unsigned int timeout=300);
     bool GetRunConfig(std::string& json_data, const std::string& name, const int version, const unsigned int timeout=300);
     bool GetDeviceConfigFromRunConfig(std::string& json_data, const int runconfig_id, const std::string& device="", const unsigned int timeout=300);
     bool GetDeviceConfigFromRunConfig(std::string& json_data, const std::string& runconfig_name, const int runconfig_version, const std::string& device="", const unsigned int timeout=300);
-    bool SendROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, bool persistent=false, int* version=nullptr, const unsigned int timestamp=0, const unsigned int timeout=300);
-    bool SendTemporaryROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, int* version=nullptr, const unsigned int timestamp=0);
-    bool SendPersistentROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, int* version=nullptr, const unsigned int timestamp=0, const unsigned int timeout=300);
+    bool SendROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, bool persistent=false, int* version=nullptr, const int64_t timestamp=0, const unsigned int timeout=300);
+    bool SendTemporaryROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, int* version=nullptr, const int64_t timestamp=0);
+    bool SendPersistentROOTplot(const std::string& plot_name, const std::string& draw_options, const std::string& json_data, int* version=nullptr, const int64_t timestamp=0, const unsigned int timeout=300);
     bool GetROOTplot(const std::string& plot_name, int& version, std::string& draw_option, std::string& json_data, std::string* timestamp=nullptr, const unsigned int timeout=300);
     bool SendPlot(Plot& plot, unsigned timeout=300);
     bool GetPlot(const std::string& name, Plot& plot, unsigned timeout=300);
